+++ src/xian/methods/finalize_block.py
@@ -5,16 +5,22 @@ from cometbft.abci.v1beta3.types_pb2 import (
     ResponseFinalizeBlock,
     ExecTxResult
 )
-from xian.utils import (
-    decode_transaction_bytes,
-    unpack_transaction,
-    get_nanotime_from_block_time,
-    convert_binary_to_hex,
-    stringify_decimals,
-    verify,
+from xian.utils.hash import (
     hash_list,
     hash_from_rewards,
     hash_from_validator_updates
+)
+from xian.utils.block import (
+    get_nanotime_from_block_time
+)
+from xian.utils.tx import (
+    verify,
+    unpack_transaction
+)
+from xian.utils.encoding import (
+    decode_transaction_bytes,
+    convert_binary_to_hex,
+    stringify_decimals
 )
 from loguru import logger
